# backend/app/features/runs/agent_router.py
"""Agent-run APIs: the tool-call timeline and approvals.

Approvals are the only way an `ask` becomes a yes. They come from the logged-in
user through this API, never from model output.
"""

import asyncio
import posixpath
import uuid
from datetime import UTC, datetime
from typing import Any, Literal

from fastapi import APIRouter, Query, Request
from pydantic import BaseModel, Field
from sqlalchemy import select

from app.api.deps import Db, client_ip
from app.core.errors import Conflict, NotFound
from app.events import bus
from app.features.audit import service as audit
from app.features.runs import service
from app.features.runs.models import Approval, FileChange, Run, ToolCall
from app.features.runs.router import PlanIn
from app.jobs import queue
from app.policy.models import Grant
from app.workspace import revert

router = APIRouter(tags=["agents"])


class ApprovalOut(BaseModel):
    id: uuid.UUID
    run_id: uuid.UUID
    tool_call_id: uuid.UUID
    kind: str  # "action" or "plan"
    status: str
    scope: str | None
    summary: str
    reason: str | None
    created_at: datetime
    decided_at: datetime | None


class ToolCallOut(BaseModel):
    id: uuid.UUID
    step: int
    position: int
    tool_name: str
    capability: str | None
    args: dict[str, Any]
    actions: list[dict[str, Any]]
    risk: str | None
    decision: str | None
    decision_reason: str | None
    status: str
    result: str | None
    result_data: dict[str, Any] | None
    is_error: bool
    started_at: datetime | None
    ended_at: datetime | None
    approval: ApprovalOut | None = None


class FileChangeOut(BaseModel):
    id: uuid.UUID
    tool_call_id: uuid.UUID | None
    op: str
    path: str
    dest_path: str | None
    reverted_at: datetime | None
    created_at: datetime


class TimelineOut(BaseModel):
    run_id: uuid.UUID
    kind: str
    status: str
    pause_requested: bool
    plan: list[dict[str, Any]] | None
    plan_version: int
    tool_calls: list[ToolCallOut]
    file_changes: list[FileChangeOut]


@router.get("/runs/{run_id}/timeline", response_model=TimelineOut)
async def timeline(run_id: uuid.UUID, db: Db) -> TimelineOut:
    run = await service.get_run(db, run_id)
    calls = list(
        await db.scalars(
            select(ToolCall)
            .where(ToolCall.run_id == run.id)
            .order_by(ToolCall.step, ToolCall.position)
        )
    )
    approvals = {
        a.tool_call_id: ApprovalOut.model_validate(a, from_attributes=True)
        for a in await db.scalars(select(Approval).where(Approval.run_id == run.id))
    }
    changes = await db.scalars(
        select(FileChange).where(FileChange.run_id == run.id).order_by(FileChange.created_at)
    )
    return TimelineOut(
        run_id=run.id,
        kind=run.kind,
        status=run.status,
        pause_requested=run.pause_requested,
        plan=run.plan,
        plan_version=run.plan_version,
        file_changes=[FileChangeOut.model_validate(c, from_attributes=True) for c in changes],
        tool_calls=[
            ToolCallOut.model_validate(c, from_attributes=True).model_copy(
                update={"approval": approvals.get(c.id)}
            )
            for c in calls
        ],
    )


class PendingApprovalOut(ApprovalOut):
    conversation_id: uuid.UUID | None
    tool_name: str


@router.get("/approvals", response_model=list[PendingApprovalOut])
async def list_approvals(
    db: Db, status: Literal["pending"] = Query("pending")
) -> list[PendingApprovalOut]:
    rows = await db.execute(
        select(Approval, Run.conversation_id, ToolCall.tool_name)
        .join(Run, Run.id == Approval.run_id)
        .join(ToolCall, ToolCall.id == Approval.tool_call_id)
        .where(Approval.status == status)
        .order_by(Approval.created_at)
    )
    return [
        PendingApprovalOut(
            **ApprovalOut.model_validate(a, from_attributes=True).model_dump(),
            conversation_id=cid,
            tool_name=name,
        )
        for a, cid, name in rows.all()
    ]


class DecisionIn(BaseModel):
    decision: Literal["approve", "deny"]
    # "run": also allow the same kind of action (same capability, same folder) for the
    # rest of this run without asking again.
    scope: Literal["once", "run"] = "once"
    reason: str | None = Field(None, max_length=500)
    # Plan reviews: the steps as edited by the user (omit to accept the plan as proposed).
    plan: PlanIn | None = None


def grant_for(call: ToolCall) -> Grant:
    """Scope an "allow for this run" approval as narrowly as practical: same capability,
    and for paths the same folder (e.g. reading projects/a/x.md grants projects/a)."""
    resource = next((a.get("resource") for a in call.actions if a.get("resource")), None)
    prefix = None
    if resource and resource != ".":
        parent = posixpath.dirname(resource.rstrip("/"))
        prefix = parent or resource
    return Grant(capability=call.capability or call.tool_name, resource_prefix=prefix)


@router.post("/approvals/{approval_id}", response_model=ApprovalOut)
async def decide(approval_id: uuid.UUID, body: DecisionIn, request: Request, db: Db) -> ApprovalOut:
    approval = await db.scalar(select(Approval).where(Approval.id == approval_id).with_for_update())
    if approval is None:
        raise NotFound("Approval not found")
    if approval.status != "pending":
        raise Conflict("This request was already answered", code="already_decided")
    run = await db.scalar(select(Run).where(Run.id == approval.run_id).with_for_update())
    call = await db.get(ToolCall, approval.tool_call_id)
    assert run is not None and call is not None
    if run.status != "waiting_approval":
        raise Conflict("The run is no longer waiting for approval", code="run_not_waiting")

    approval.status = "approved" if body.decision == "approve" else "denied"
    approval.scope = body.scope if body.decision == "approve" else None
    approval.reason = body.reason
    approval.decided_at = datetime.now(UTC)
    if approval.kind == "plan":
        approval.scope = "once" if body.decision == "approve" else None
        if body.decision == "approve" and body.plan is not None:
            call.args = {"steps": [s.model_dump() for s in body.plan.steps]}
    elif body.decision == "approve" and body.scope == "run":
        run.grants = [*run.grants, grant_for(call).model_dump()]
    audit.record(
        db,
        f"approval.{approval.status}",
        target_type="tool_call",
        target_id=call.id,
        ip=client_ip(request),
        details={"run_id": str(run.id), "summary": approval.summary, "scope": body.scope},
    )
    await queue.enqueue(db, "run.execute", {"run_id": str(run.id)}, lane="interactive")
    await bus.publish_run_event(
        run.id, "approval.resolved", {"approval_id": str(approval.id), "status": approval.status}
    )
    await service.set_status(db, run, "queued")  # commits
    return ApprovalOut.model_validate(approval, from_attributes=True)


class RevertOut(BaseModel):
    message: str
    change: FileChangeOut


@router.post("/runs/{run_id}/files/{change_id}/revert", response_model=RevertOut)
async def revert_change(
    run_id: uuid.UUID, change_id: uuid.UUID, request: Request, db: Db, force: bool = False
) -> RevertOut:
    """Undo one file change made by an agent in this run."""
    change = await db.get(FileChange, change_id)
    if change is None or change.run_id != run_id:
        raise NotFound("Change not found")
    message = await asyncio.to_thread(revert.revert, change, force=force)
    audit.record(
        db,
        "files.revert",
        target_type="file_change",
        target_id=change.id,
        ip=client_ip(request),
        details={"message": message, "force": force},
    )
    await db.commit()
    return RevertOut(
        message=message, change=FileChangeOut.model_validate(change, from_attributes=True)
    )