# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anemo",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "description": "Run history, newest first. Page with `before` = the last item's created_at.",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "agent",
                "chat",
                "all"
              ],
              "type": "string",
              "default": "agent",
              "title": "Kind"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "waiting",
                "completed",
                "failed",
                "cancelled",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in the request text",
              "title": "Q"
            },
            "description": "Search in the request text"
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only runs created before this time",
              "title": "Before"
            },
            "description": "Only runs created before this time"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunListItem"
                  },
                  "title": "Response List Runs Api Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/timeline": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Timeline",
        "operationId": "timeline_api_runs__run_id__timeline_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/approvals": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Approvals",
        "operationId": "list_approvals_api_approvals_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "const": "pending",
              "type": "string",
              "default": "pending",
              "title": "Status"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PendingApprovalOut"
                  },
                  "title": "Response List Approvals Api Approvals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/approvals/{approval_id}": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Decide",
        "operationId": "decide_api_approvals__approval_id__post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/files/{change_id}/revert": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Revert Change",
        "description": "Undo one file change made by an agent in this run.",
        "operationId": "revert_change_api_runs__run_id__files__change_id__revert_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Change Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevertOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionIn": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "deny"
            ],
            "title": "Decision"
          },
          "scope": {
            "type": "string",
            "enum": [
              "once",
              "run"
            ],
            "title": "Scope",
            "default": "once"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanIn"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecisionIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PendingApprovalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "tool_call_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tool Call Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "run_id",
          "tool_call_id",
          "kind",
          "status",
          "scope",
          "summary",
          "reason",
          "created_at",
          "decided_at",
          "conversation_id",
          "tool_name"
        ],
        "title": "PendingApprovalOut"
      },
...
      "PlanIn": {
        "properties": {
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PlanStepIn"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "steps"
        ],
        "title": "PlanIn"
      },
...
      "RevertOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "change": {
            "$ref": "#/components/schemas/FileChangeOut"
          }
        },
        "type": "object",
        "required": [
          "message",
          "change"
        ],
        "title": "RevertOut"
      },
...
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "assistant_message_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistant Message Id"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Name"
          },
          "request": {
            "type": "string",
            "title": "Request"
          },
          "pause_requested": {
            "type": "boolean",
            "title": "Pause Requested"
          },
          "plan_version": {
            "type": "integer",
            "title": "Plan Version"
          },
          "totals": {
            "$ref": "#/components/schemas/RunTotals"
          },
          "limits": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limits"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "conversation_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Title"
          },
          "model_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Label"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "step",
          "status",
          "conversation_id",
          "assistant_message_id",
          "model_id",
          "profile_id",
          "request",
          "pause_requested",
          "plan_version",
          "totals",
          "error",
          "created_at",
          "started_at",
          "ended_at"
        ],
        "title": "RunDetail"
      },
      "RunListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "assistant_message_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistant Message Id"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Name"
          },
          "request": {
            "type": "string",
            "title": "Request"
          },
          "pause_requested": {
            "type": "boolean",
            "title": "Pause Requested"
          },
          "plan_version": {
            "type": "integer",
            "title": "Plan Version"
          },
          "totals": {
            "$ref": "#/components/schemas/RunTotals"
          },
          "limits": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limits"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "conversation_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Title"
          },
          "model_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Label"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "step",
          "status",
          "conversation_id",
          "assistant_message_id",
          "model_id",
          "profile_id",
          "request",
          "pause_requested",
          "plan_version",
          "totals",
          "error",
          "created_at",
          "started_at",
          "ended_at"
        ],
        "title": "RunListItem"
      },
...
      "TimelineOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "pause_requested": {
            "type": "boolean",
            "title": "Pause Requested"
          },
          "plan": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan"
          },
          "plan_version": {
            "type": "integer",
            "title": "Plan Version"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCallOut"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "file_changes": {
            "items": {
              "$ref": "#/components/schemas/FileChangeOut"
            },
            "type": "array",
            "title": "File Changes"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "kind",
          "status",
          "pause_requested",
          "plan",
          "plan_version",
          "tool_calls",
          "file_changes"
        ],
        "title": "TimelineOut"
      },
...
    }
  }
}